Basic support for different HTTP clients
