+++ lib/ApiRequestor.php
@@ -8,6 +8,8 @@ class ApiRequestor
 
     private $_apiBase;
 
+    private static $_httpClient;
+
     public function __construct($apiKey = null, $apiBase = null)
     {
         $this->_apiKey = $apiKey;
@@ -17,25 +19,10 @@ class ApiRequestor
         $this->_apiBase = $apiBase;
     }
 
-    /**
-     * @param string|mixed $value A string to UTF8-encode.
-     *
-     * @return string|mixed The UTF8-encoded string, or the object passed in if
-     *    it wasn't a string.
-     */
-    public static function utf8($value)
-    {
-        if (is_string($value) && mb_detect_encoding($value, "UTF-8", true) != "UTF-8") {
-            return utf8_encode($value);
-        } else {
-            return $value;
-        }
-    }
-
     private static function _encodeObjects($d)
     {
         if ($d instanceof ApiResource) {
-            return self::utf8($d->id);
+            return Util\Util::utf8($d->id);
         } elseif ($d === true) {
             return 'true';
         } elseif ($d === false) {
@@ -47,42 +34,8 @@ class ApiRequestor
             }
             return $res;
         } else {
-            return self::utf8($d);
+            return Util\Util::utf8($d);
         }
-    }
-
-    /**
-     * @param array $arr An map of param keys to values.
-     * @param string|null $prefix (It doesn't look like we ever use $prefix...)
-     *
-     * @return string A querystring, essentially.
-     */
-    public static function encode($arr, $prefix = null)
-    {
-        if (!is_array($arr)) {
-            return $arr;
-        }
-
-        $r = array();
-        foreach ($arr as $k => $v) {
-            if (is_null($v)) {
-                continue;
-            }
-
-            if ($prefix && $k && !is_int($k)) {
-                $k = $prefix."[".$k."]";
-            } elseif ($prefix) {
-                $k = $prefix."[]";
-            }
-
-            if (is_array($v)) {
-                $r[] = self::encode($v, $k, true);
-            } else {
-                $r[] = urlencode($k)."=".urlencode($v);
-            }
-        }
-
-        return implode("&", $r);
     }
 
     /**
@@ -209,7 +162,7 @@ class ApiRequestor
             $rawHeaders[] = $header . ': ' . $value;
         }
 
-        list($rbody, $rcode) = $this->_curlRequest(
+        list($rbody, $rcode) = $this->httpClient()->request(
             $method,
             $absUrl,
             $rawHeaders,
@@ -258,116 +211,16 @@ class ApiRequestor
         return $resp;
     }
 
-    private function _curlRequest($method, $absUrl, $headers, $params, $hasFile)
+    public static function setHttpClient($client)
     {
-        $curl = curl_init();
-        $method = strtolower($method);
-        $opts = array();
-        if ($method == 'get') {
-            if ($hasFile) {
-                throw new Error\Api(
-                    "Issuing a GET request with a file parameter"
-                );
-            }
-            $opts[CURLOPT_HTTPGET] = 1;
-            if (count($params) > 0) {
-                $encoded = self::encode($params);
-                $absUrl = "$absUrl?$encoded";
-            }
-        } elseif ($method == 'post') {
-            $opts[CURLOPT_POST] = 1;
-            $opts[CURLOPT_POSTFIELDS] = $hasFile ? $params : self::encode($params);
-        } elseif ($method == 'delete') {
-            $opts[CURLOPT_CUSTOMREQUEST] = 'DELETE';
-            if (count($params) > 0) {
-                $encoded = self::encode($params);
-                $absUrl = "$absUrl?$encoded";
-            }
-        } else {
-            throw new Error\Api("Unrecognized method $method");
-        }
-
-        $absUrl = self::utf8($absUrl);
-        $opts[CURLOPT_URL] = $absUrl;
-        $opts[CURLOPT_RETURNTRANSFER] = true;
-        $opts[CURLOPT_CONNECTTIMEOUT] = 30;
-        $opts[CURLOPT_TIMEOUT] = 80;
-        $opts[CURLOPT_RETURNTRANSFER] = true;
-        $opts[CURLOPT_HTTPHEADER] = $headers;
-        if (!Stripe::$verifySslCerts) {
-            $opts[CURLOPT_SSL_VERIFYPEER] = false;
-        }
-
-        curl_setopt_array($curl, $opts);
-        $rbody = curl_exec($curl);
-
-        if (!defined('CURLE_SSL_CACERT_BADFILE')) {
-            define('CURLE_SSL_CACERT_BADFILE', 77);  // constant not defined in PHP
-        }
-
-        $errno = curl_errno($curl);
-        if ($errno == CURLE_SSL_CACERT ||
-            $errno == CURLE_SSL_PEER_CERTIFICATE ||
-            $errno == CURLE_SSL_CACERT_BADFILE
-        ) {
-            array_push(
-                $headers,
-                'X-Stripe-Client-Info: {"ca":"using Stripe-supplied CA bundle"}'
-            );
-            $cert = $this->caBundle();
-            curl_setopt($curl, CURLOPT_HTTPHEADER, $headers);
-            curl_setopt($curl, CURLOPT_CAINFO, $cert);
-            $rbody = curl_exec($curl);
-        }
-
-        if ($rbody === false) {
-            $errno = curl_errno($curl);
-            $message = curl_error($curl);
-            curl_close($curl);
-            $this->handleCurlError($errno, $message);
-        }
-
-        $rcode = curl_getinfo($curl, CURLINFO_HTTP_CODE);
-        curl_close($curl);
-        return array($rbody, $rcode);
+        self::$_httpClient = $client;
     }
 
-    /**
-     * @param number $errno
-     * @param string $message
-     * @throws ApiConnectionError
-     */
-    public function handleCurlError($errno, $message)
+    private function httpClient()
     {
-        $apiBase = $this->_apiBase;
-        switch ($errno) {
-            case CURLE_COULDNT_CONNECT:
-            case CURLE_COULDNT_RESOLVE_HOST:
-            case CURLE_OPERATION_TIMEOUTED:
-                $msg = "Could not connect to Stripe ($apiBase).  Please check your "
-                 . "internet connection and try again.  If this problem persists, "
-                 . "you should check Stripe's service status at "
-                 . "https://twitter.com/stripestatus, or";
-                break;
-            case CURLE_SSL_CACERT:
-            case CURLE_SSL_PEER_CERTIFICATE:
-                $msg = "Could not verify Stripe's SSL certificate.  Please make sure "
-                 . "that your network is not intercepting certificates.  "
-                 . "(Try going to $apiBase in your browser.)  "
-                 . "If this problem persists,";
-                break;
-            default:
-                $msg = "Unexpected error communicating with Stripe.  "
-                 . "If this problem persists,";
+        if (!self::$_httpClient) {
+            self::$_httpClient = HttpClient\CurlClient::instance();
         }
-        $msg .= " let us know at support@stripe.com.";
-
-        $msg .= "\n\n(Network error [errno $errno]: $message)";
-        throw new Error\ApiConnection($msg);
-    }
-
-    private function caBundle()
-    {
-        return dirname(__FILE__) . '/../data/ca-certificates.crt';
+        return self::$_httpClient;
     }
 }
