+++ tests/UtilTest.php
@@ -27,4 +27,19 @@ class UtilTest extends TestCase
         $customer = self::createTestCustomer();
         $this->assertTrue(array_key_exists("id", $customer->__toArray(true)));
     }
+
+    public function testUtf8()
+    {
+        // UTF-8 string
+        $x = "\xc3\xa9";
+        $this->assertSame(Util\Util::utf8($x), $x);
+
+        // Latin-1 string
+        $x = "\xe9";
+        $this->assertSame(Util\Util::utf8($x), "\xc3\xa9");
+
+        // Not a string
+        $x = true;
+        $this->assertSame(Util\Util::utf8($x), $x);
+    }
 }
