+++ tests/bootstrap.no_autoload.php
@@ -2,3 +2,4 @@
 
 require_once __DIR__ . '/../init.php';
 require_once __DIR__ . '/TestCase.php';
+require_once __DIR__ . '/MockHttpClient.php';
