+++ tests/bootstrap.php
@@ -1,3 +1,4 @@
 <?php
 
 require_once __DIR__ . '/TestCase.php';
+require_once __DIR__ . '/MockHttpClient.php';
